ACTUALLY fix bug where campuses weren't displaying
Checking campus didn't work because the length got set to the string
length for the error messages; checking campus.data didn't work because
it was always undefined (data coming from response.data).

What I wanted to check was if campus returned a string, so setting the
condition to typeof !== "string" works because normally it returns an
array. Yaaaaay /o/