Update delete button function
Original code interacted with dummy data. The "correct" way would be to
fetch the new response.data after making the DELETE request.

Please note that I've changed the dependency of the function to
re-render on setCampus, forcing a refresh for changes to update. It's
better than a memory leak...

(breaks) and fixes #10